+++ src/components/index.tsx
@@ -1,0 +1,23 @@
+import { AddPostMobile } from "./AddPostMobile";
+import { AddQuestionModal } from "./AddQuestionModal";
+import { CustomSnackbar } from "./CustomSnackbar";
+import { DiscussionCard } from "./DiscussionCard";
+import { DiscussionCardsSection } from "./DiscussionCardsSection";
+import { FullscreenLoader } from "./FullscreenLoader";
+import { Navbar } from "./Navbar";
+import { RequireAuth } from "./RequireAuth";
+import { Sidenav } from "./Sidenav";
+import { TrendingSection } from "./TrendingSection";
+
+export {
+  AddPostMobile,
+  AddQuestionModal,
+  CustomSnackbar,
+  DiscussionCard,
+  DiscussionCardsSection,
+  FullscreenLoader,
+  Navbar,
+  RequireAuth,
+  Sidenav,
+  TrendingSection,
+};
