+++ src/pages/index.tsx
@@ -1,0 +1,6 @@
+import { Feed } from "./feed";
+import { LandingPage } from "./LandingPage";
+import { SignIn } from "./Signin";
+import { SignUp } from "./SignUp";
+
+export { Feed, LandingPage, SignIn, SignUp };
